chore(julia): create a 2nd version of complex_point with range
This commit adds a 2nd version of the complex point calculation in the `generate_julia_set` function. It introduces a new variable `range` and modifies the calculation of `complex_point`. The previous implementation using `descriptor.to_complex` is commented out for now. Further discussion with the group is needed to finalize this change.