feat(cli): Add save flag to CLI client arguments
Added an optional flag to the CLI client arguments that allows users to specify whether they want to save the generated image to a file. The default value for this flag is set to false. When processing a fragment task, if the save flag is enabled, the generated fractal image will be saved to a file. Additionally, if the open flag is enabled, the saved image will be opened after saving.